% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata keywords to a file using exifTools}
\usage{
add_metadata(
  filePath,
  ...,
  exifPath = "../../../exiftool-12.65/exiftool.exe",
  addToFile = TRUE
)
}
\arguments{
\item{filePath}{Which file to add metadata to?}

\item{...}{Anything else to add as keywords beyond the the params and file
code. Include as "name" = value.}

\item{exifPath}{The filepath to the ExifTools executable.}

\item{addToFile}{BINARY: If FALSE, the keywords are invisibly returned, and
the call to exifTools is skipped.}
}
\description{
A way to document which params were used for a figure, and to
track the file and latest git commit. Only tested with the windows
executable.
}
