% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_magnitude.R
\name{calc_magnitude}
\alias{calc_magnitude}
\title{Return the magnitude of a number}
\usage{
calc_magnitude(x)
}
\arguments{
\item{x}{The number.}
}
\description{
Magnitudes are considered powers of 10, so 1:9.999... are all
magnitude 0, 10-99.999... are magnitude 1, 0.1:0.999... are -1, etc. The
sign of the number makes no difference.
}
