% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Return the mode average - most common value/s}
\usage{
calc_mode(x)
}
\arguments{
\item{x}{The data to find the mode of.}
}
\description{
R has built in mean and median functions, but not mode.
}
