% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_quantile_range.R
\name{calc_quantile_range}
\alias{calc_quantile_range}
\title{Find custom interquantile ranges}
\usage{
calc_quantile_range(x, lower = 0.1, upper = 0.9, ...)
}
\arguments{
\item{x}{numeric vector whose sample quantiles are wanted, or an
    object of a class for which a method has been defined (see also
    \sQuote{details}). \code{\link{NA}} and \code{NaN} values are not
    allowed in numeric vectors unless \code{na.rm} is \code{TRUE}.}

\item{lower}{The lower probability.}

\item{upper}{The upper probability.}

\item{...}{Any other arguments that can be accepted by \code{\link[stats:quantile]{stats::quantile()}},
except 'probs'.}
}
\description{
Equivalent to the IQR, but not restricted to the 25th & 75th
percentiles. Simply calls \code{\link[stats:quantile]{stats::quantile()}}, and takes the lower value
from the upper value.
}
