% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2.R
\name{cat3}
\alias{cat3}
\title{Concatenate a variable}
\usage{
cat3(..., show = TRUE)
}
\arguments{
\item{...}{The variable/s that should be displayed. The values will be
collapsed down with a space between them and then shown using \code{cat()}.
Not tested with complex data types - use with caution! If you want the
name of the variable and the variable, use \code{cat2()}.}

\item{show}{logical If FALSE, the function does nothing; it is useful if
there are multiple \code{cat3()} calls throughout; set a variable at the
beginning and use that as the 'show' argument in each \code{cat3()} call.
Simply changing that to FALSE allows the function to be run without all
the calls, but the calls can remain in place if further \code{cat3()} use is
likely.}
}
\description{
This function is simply \code{cat()} but adds a newline \verb{\\n} before
and after to quicken things up.
}
