% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat4.R
\name{cat4}
\alias{cat4}
\title{Print out a variable's name, value and the location}
\usage{
cat4(x, show = TRUE, skipCatCall = TRUE)
}
\arguments{
\item{x}{The variable to show.}

\item{show}{LOGICAL: If FALSE, the function just returns the input but does
not print anything out. This behaviour is helpful as it allows multiple
\code{\link[=cat2]{cat2()}} calls to be added throughout a function, and a single variable
at the start can be used to turn the printing on for debugging, or off
for use. Useful when it is likely that further debugging will be needed.}

\item{skipCatCall}{When nested in another function (for example \code{\link[=cat4]{cat4()}}),
should the call to that function (e.g.\code{cat4(x)}) be included in the print
out of where \code{\link[=which_line]{which_line()}} was called from? TRUE by default.}
}
\description{
This function combines \code{\link[=which_line]{which_line()}} and \code{\link[=cat2]{cat2()}}. It
therefore shows the value and name of a variable, and the location of it.
This is useful for helping to debug when writing functions.
}
