% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat5.R
\name{cat5}
\alias{cat5}
\title{Print out a variable's name, class and value}
\usage{
cat5(x, show = TRUE, name = NULL)
}
\arguments{
\item{x}{The variable that should be displayed. It's name and value will be
shown. Lists are fed to \code{\link[=cat_list]{cat_list()}}.

If you just want the variable (and not the name and class), stick with
\code{\link[=cat]{cat()}}, or use \code{\link[=cat3]{cat3()}}, which basically adds a newline \verb{\\n} before and
after x; if just the name and value, use \code{\link[=cat2]{cat2()}}.

\strong{Note:} This function has not been tested with complex data types - use
with caution!}

\item{show}{LOGICAL: If FALSE, the function just returns the input but does
not print anything out. This behaviour is helpful as it allows multiple
\code{\link[=cat5]{cat5()}} calls to be added throughout a function, and a single variable
at the start can be used to turn the printing on for debugging, or off
for use. Useful when it is likely that further debugging will be needed.}

\item{name}{"string": Should an alternative name be used? By default, the
name of 'x' is used. However, if \code{\link[=cat5]{cat5()}} is nested, this will likely
just end up as "x" for every variable, so the true name can be entered
here.}
}
\description{
Using \code{\link[=cat5]{cat5()}} prints out a variable, it's class and it's name.

This function is useful because I don't know how to debug "correctly".
For example, when developing functions, it can be hard to know where the
code is going wrong, particularly if there are no true errors.  However,
the returned value can still be incorrect, but it is not clear at which
stage a variable has which value.

Sprinkling the \code{\link[=cat5]{cat5()}} function at key locations in functions allows us
to see what is happening where within a function.
}
