% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_path.R
\name{cat_path}
\alias{cat_path}
\title{Concatenate input into a filepath}
\usage{
cat_path(...)
}
\arguments{
\item{...}{The parts of the file path. Each part will be separated with a
"/". Can handle vectors or lists.}
}
\description{
A very simple function that simply collapses the input down,
separating each part with a "/" to create a filepath. It is very similar
to file.path, but just my (slower) implementation.
}
