% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_progress.R
\name{cat_progress}
\alias{cat_progress}
\title{A simple progress bar}
\usage{
cat_progress(progress, total)
}
\arguments{
\item{progress}{How many steps have been completed?}

\item{total}{How many steps are there in total?}
}
\description{
This function displays the progress on the console. It is a
simple function that is just here as a simple test; the intention is to
develop it further.
}
\examples{
  # Watch it in progress
  for (ii in 1:15) {
    cat_progress(ii, 15)
    Sys.sleep(0.2)
  }

  # Staggered output to observe in time
  for (ii in 1:8) {
    cat_progress(ii, 15)
    cat("\n")
  }

  # More "realistic" use
  # fake function
  tst_fun <- function(x) {
    # pretend we're doing something more involve & slower!
    yTime <- sample(c(0:100), 10, replace = TRUE) / 100
    Sys.sleep(yTime)
    y <- x^3
    return(y)
  }

  # fake data
  xx <- sample(1:1000, 15)
  tt <- rep("", length(xx)) # preallocate

  # loop
  for (ii in seq_along(tt)) {
    tt[ii] <- tst_fun(ii)
    cat_progress(ii, length(tt))
  }

}
