% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_list_depth.R
\name{count_list_depth}
\alias{count_list_depth}
\title{Count how deep nested lists go}
\usage{
count_list_depth(list, depth = 0)
}
\arguments{
\item{list}{The list to counted nested list levels within.}

\item{depth}{Used to track the depth in recursive calls.}
}
\description{
Lists can be nested. This simply counts how many nested levels
there are. This code is largely copied from Spacedman's stackoverflow
answer here: \url{https://stackoverflow.com/questions/13432863/determine-level-of-nesting-in-r}.
}
