% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lean_bib_file.R
\name{create_lean_bib}
\alias{create_lean_bib}
\title{Create a leaner .bib file}
\usage{
create_lean_bib(inputFile, inputBib, outputBib)
}
\arguments{
\item{inputFile}{"string": path to an ".rmd" or ".qmd" file to check for
citations in}

\item{inputBib}{"string": path to the .bib file used for the inputFile}

\item{outputBib}{"string": path location for the new, leaner .bib file}
}
\description{
Instead of sharing a full Zotero library, this function
creates a leaner .bib file that only stores the citations that were used.
It has not been extensively tested, but it worked okay for use with the
phd01 manuscript. The code was modified from this StackOverflow answer:
\url{https://stackoverflow.com/questions/65306015/in-rmarkdown-is-there-a-way-to-create-a-bib-file-for-only-those-keys-cited-in}.
It assumes that all references begin with \code{"@"} (e.g. \verb{@Saun2022}). It
may therefore grab any email addresses in there, so it is not foolproof.
}
\examples{
\dontrun{
    create_lean_bib("path/to/manuscript/phd01_manuscript.Rmd",
                    "path/to/existing/bibliography/zotero.bib",
                    "path/to/new/lean/bib/tc-2022-94_bibliography.bib")
}

}
