% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fresh.R
\name{fresh}
\alias{fresh}
\title{Clear the console, environment, custom functions, and/or plots.}
\usage{
fresh(toKeep = "f")
}
\arguments{
\item{toKeep}{"string": Which of the four options should be kept?
\itemize{
\item f    functions
\item v    variables/values
\item p    plots
\item c    console
\item e    environment (i.e. functions & variables)
}}
}
\description{
This function is potentially controversial! However, it helps
with the way I approach R, whether that is rightly or wrongly. At the
start of a script / notebook for example, it can be helpful to clear
everything and reset to see that things are working and not relying on
something you didn't realise was in the memory. It is not the same as
restarting R or closing RStudio, but I haven't run into an instance where
it is actually an issue.
}
\examples{
\dontrun{
  # Clear the variables, plots, and console, but keep functions:
  fresh()     # default behaviour

  # Clear the environment and console, but keep the plots:
  fresh("p")

  # Clear everything
  fresh("")
}

}
