% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_di_matrix.R
\name{get_di_matrix}
\alias{get_di_matrix}
\title{Transforms a frequency interaction sociomatrix (valued data) into a dichotomized 1/0 matrix}
\usage{
get_di_matrix(m, type = "wl")
}
\arguments{
\item{m}{A matrix with individuals ordered identically in rows and columns.}

\item{type}{Determines the type of dichotomized matrix to be returned.
\strong{\code{type}="wl"} is the default which returns a win-loss matrix
with a '1' representing a consistent winner and a '0' representing a
consistent loser for each dyad of the matrix. A consistent winner is
defined as being the individual in each dyad that has absolutely more
wins than defeats.   In the default condition if competitors have the
same number of wins each, they both receive a 0.
If \strong{\code{type}="wlties"} the default dichotomized win-loss
matrix will be returned but it will also return 0.5 into cells for tied
relationships.}
}
\value{
A dichotomized win/loss or presence/absence matrix.
}
\description{
Transforms a frequency interaction sociomatrix (valued data) into a dichotomized 1/0 matrix
}
\section{References}{

Appleby, M. C. 1983. The probability of linearity in hierarchies.
Animal Behaviour, 31, 600-608.
}

\examples{
get_di_matrix(bonobos)
get_di_matrix(caribou)
}
