% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extension.R
\name{get_extension}
\alias{get_extension}
\title{Extract the file extension from a filename}
\usage{
get_extension(filename)
}
\arguments{
\item{filename}{The filename we want the extension from.}
}
\description{
A very simple function that takes whatever is to the right of
the last "." in a filename and considers that as the file extension.
}
