% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_git.R
\name{get_latest_git}
\alias{get_latest_git}
\title{Print the information of the latest git commit}
\usage{
get_latest_git(display = TRUE, incGitCommit = FALSE)
}
\arguments{
\item{display}{BINARY: Should the output be displayed? Usually, this is the
whole purpose of the function, but sometimes the output should be
suppressed, for example if adding the information to a list of data that
will then be saved in the "Data/" folder.}

\item{incGitCommit}{BINARY: Should the "gitcommit" object be returned? Most
of the data has already been "repurposed" so this is usually unnecessary,
plus some functions do not know how to handle a "gitcommit" object. It is
therefore FALSE by default. Set to TRUE if you really want it.}
}
\description{
When we use quarto notebooks, and render them to a pdf, we
don't necessarily know where the codebase was at for the pdf. In future,
we may want to go back to different version, but can only guess on the
date of rendering. This function is useful for such an instance because
it allows us to add the latest git commit before rendering so we have a
better idea. Requires the \code{git2r} package.
}
