% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Prepare metadata about a file}
\usage{
get_metadata(..., format = "list", incGit = TRUE, incSession = TRUE)
}
\arguments{
\item{...}{Any additional information that needs to be included beyond the
params and file code? Include as '"name" =  value'.}

\item{format}{Which format do you want the metadata in? Options are "pdf",
"NetCDF", "print", "named", or "yaml" / "json" / "list" (all the same).
"list" is the default and the fallback if anything except one of these
strings is entered.

Options (examples of formatting); names are cASe inSENSItIve.

\if{html}{\out{<div class="sourceCode">}}\preformatted{ "pdf"                      "-Keywords=month:Dec"
 "NetCDF"                   "month=Dec"
 "print"                    "month : Dec"
 "named"                    named vector: "month" = "Dec"
 "yaml" / "json" / "list"   list("month" = "Dec")
}\if{html}{\out{</div>}}}

\item{incGit}{BINARY: Should the full output from \code{\link[=get_latest_git]{get_latest_git()}} be
included? Even if FALSE, the hash of the last commit is still included in
the file code.}

\item{incSession}{BINARY: Should the output from \code{\link[=sessionInfo]{sessionInfo()}} be
included? This is currently not implemented as it looks more complex than
I had anticipated.}
}
\description{
Get the information necessary for documenting how a file /
dataset was created. It includes a time and date for when it was called,
the filecode (see \code{\link[=get_file_codes]{get_file_codes()}}) of  the file it was called from;
any 'params' if called from a .qmd file, and (optionaly) information on
the last git commit (requires \href{https://docs.ropensci.org/git2r/}{git2r}),
the output from \code{\link[=sessionInfo]{sessionInfo()}}, and any additional information that is
entered as named arguments (e.g. '"extraMetadata" = "useful
information"').
}
