% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sessionInfo.R
\name{get_sessionInfo}
\alias{get_sessionInfo}
\title{Simplify sessionInfo() for json or yaml}
\usage{
get_sessionInfo()
}
\description{
The default output from \code{\link[=sessionInfo]{sessionInfo()}} doesn't play nicely
when trying to add metadata to a list that is saved in a json or yaml
format. This function takes the output, extracts the interesting parts
(very subjectively!) and formats them as a named list. Mainly designed to
be called by get_metadata as the normal call to \code{\link[=sessionInfo]{sessionInfo()}} is better
for printing it out.
}
