% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sentence_case.R
\name{make_sentence_case}
\alias{make_sentence_case}
\title{Convert a string into sentence case.}
\usage{
make_sentence_case(x)
}
\arguments{
\item{x}{"string": The string to turn into sentence case.}
}
\description{
If a string arrives as "hello" or "HELLO" (or e.g. "heLLo"),
this function changes it to "Hello". I took the code directly from:
\url{https://stackoverflow.com/questions/43592419/convert-to-sentence-case-in-r}.
I do not really understand how it works.
}
