% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_title_case.R
\name{make_title_case}
\alias{make_title_case}
\title{Convert a string into Title Case.}
\usage{
make_title_case(x)
}
\arguments{
\item{x}{"string": The string to turn into title case. Can be a vector.}
}
\description{
If a string arrives as "hello uSEr" or "HELLO user", etc, this
function changes it to "Hello User". Here, "Title Case" means the first
letter of each word is capitalised. Words are defined by the space
between them.
}
