% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_matrix.R
\name{mirror_matrix_horizontally}
\alias{mirror_matrix_horizontally}
\title{Mirrors a matrix so the left is right and the right is left}
\usage{
mirror_matrix_horizontally(x)
}
\arguments{
\item{x}{matrix: The matrix to mirror.}
}
\description{
Sometime the matrix is not in the orientation we want. Use
this function if you want to mirror it across a vertical line down the
centre of the matrix, so that the columns on the left move to the right,
and the columns on the right move to the left.
}
