% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirror_matrix.R
\name{mirror_matrix_vertically}
\alias{mirror_matrix_vertically}
\title{Mirrors a matrix so the top is bottom and the bottom is top}
\usage{
mirror_matrix_vertically(x)
}
\arguments{
\item{x}{matrix: The matrix to mirror.}
}
\description{
Sometime the matrix is not in the orientation we want. Use
this function if you want to mirror it across a horizontal line through
the centre of the matrix, so that the rows on the top move to the bottom,
and the rows on the bottom move to the top
}
