% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_line.R
\name{print_line}
\alias{print_line}
\title{Print a line to the console to help delineate output}
\usage{
print_line(lineType = "=", width = 22)
}
\arguments{
\item{lineType}{"string": The symbol/s that will be printed to create a
line across the console.}

\item{width}{numeric : How many times should the above be repeated? Usually
unnecessary, but change it if screen is particularly wide or narrow.}
}
\description{
Useful when working interactively in RStudio, and can be
particularly helpful to group output to the console or when testing.
}
\examples{
  print_line()
  print_line("-", 10)
  print_line("+-", 25)

}
