% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_on_saving.R
\name{print_on_saving}
\alias{print_on_saving}
\title{Let the user know the save didn't throw an error & print relevant info}
\usage{
print_on_saving(filename = "")
}
\arguments{
\item{fileName}{"string": Where should it have been saved?}
}
\description{
Just used interactively to verify that something has been
saved. We also show the filename and path, and the last git commit's SHA.
Make sure to place this function after the save call, because it doesn't
actually check anything at all - it just assumes that if it has been
called, the save went ahead okay.
}
