% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.R
\name{save_list}
\alias{save_list}
\title{A very simple wrapper for saving a list with \code{jsonlite}}
\usage{
save_list(x, filename)
}
\arguments{
\item{x}{The list to save.}

\item{filename}{The filename (and path relative to the working directory)
indicating where to save the list. Must included the file extension.}
}
\description{
This function will be expanded to also accommodate yaml, rds
and rData output, and different json options. It currently just converts
an R list to a json format using \code{jsonlite::serialize()}, and saves it
via \code{\link[=writeLines]{writeLines()}}. It uses the default values of 'pretty' = TRUE.
}
