% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sift.R
\name{sift}
\alias{sift}
\title{Sift through a dataset and return only data matching a defined condition.}
\usage{
sift(x, condition, inConditions = NA, noWarning = FALSE)
}
\arguments{
\item{x}{Which data needs to be sifted through?}

\item{condition}{"string": Which condition should be met? There are 3
necessary parts to this definition, which must each be separated with a
single space.

\if{html}{\out{<div class="sourceCode">}}\preformatted{These 3 parts are:
    ColumnName            Which column is the filter based on? Must be an exact match.
    Operator              Which condition?
    Value                 Which value?

The possible operators are:
    ">"                   Greater than
    "<"                   Less than
    "in"                  in a range / vector
    "!in" or "notIn"      not in a range / vector
    "="                   equals / the same as
    "=!"                  doesn't equal / is different to
}\if{html}{\out{</div>}}

If the "in" or "!in" Value is more than just an "a:b" vector, it is
necessary to supply the vector via the "inConditions" argument, and use
"ColumnName in vector", or "ColumnName !in vector" as the condition. See
the examples.}

\item{inConditions}{vector: a vector of values to look for; requires the
condition to be "ColumnName in vector" or "ColumnName !in vector".}

\item{noWarning}{binary: TRUE suppresses a warning when no data is returned
by the condition.}
}
\description{
Used to return only the rows of a data.frame (or matrix) which
match a certain condition. An example would be to return all the rows
from a matrix where the column containing the "January" value is greater
than "x". See the examples. This function is simply indexing into the
data, but wrapping it up in a simpler, more approachable syntax (to me!).
For multiple conditions, just chain the \code{sift()} calls together (see
example x4).
}
\examples{
    x  <- data.frame("Year" = 1961:2000,
                     "Melt" = sample(0:100, 40))
    x1 <- sift(x, "Year > 1978")
    x2 <- sift(x, "Year in 1978:1984")
    x3 <- sift(x, "Year in vector", c(1978:1984, 1987, 1989))
    x4 <- sift(x, "Melt > 50") |> sift("Year < 1980")

}
