% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_folder.R
\name{source_folder}
\alias{source_folder}
\title{Source every file in a folder}
\usage{
source_folder(folderName, verbose = FALSE, showWarnings = TRUE)
}
\arguments{
\item{folderName}{Where are the files to source?}

\item{verbose}{logical. Print details to the console?}

\item{showWarnings}{logical. Should a warning be thrown if no files found?}
}
\description{
Useful when loading all custom functions in a project for
example. Simply finds any file with an ".R" extension, and calls it with
\code{\link[=source]{source()}}.

Almost identical to this Stackoverflow answer:
\url{https://stackoverflow.com/a/20083589}
}
