% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{start_timer}
\alias{start_timer}
\title{Starts the timer}
\usage{
start_timer()
}
\description{
This function and \code{end_timer()} are mainly because I can never
remember the syntax/name of \code{proc.time()}. Simply call \code{start_timer()}
before the process you are timing begins, and then call \code{end_timer()}
afterwards, feeding it with the output of \code{start_timer()}.
}
\examples{
\dontrun{
  start <- start_timer()
  Sys.sleep(0.2) # do code!
  end_timer(start)
}
}
