% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test that the package has loaded correctly and the functions are accessible}
\usage{
test_function(name = "Dominic")
}
\arguments{
\item{name}{"string": user name}
}
\description{
Test that the package has loaded correctly and the functions are accessible
}
