% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttri.R
\name{ttri}
\alias{ttri}
\title{Get Triangle Transitivity of Sociomatrix}
\usage{
ttri(m)
}
\arguments{
\item{m}{A frequency binary win-loss matrix}
}
\value{
Pt and t.tri
}
\description{
Get Triangle Transitivity of Sociomatrix
}
\section{Further details}{

Algorithm described in D. Shizuka and D. B. McDonald, 2012,
A social network perspective on transitivity and linearity
in dominance hierarchies.  Animal Behaviour.
DOI:10.1016/j.anbehav.2012.01.011
Code adapted from original code by Dai Shikua - see
http://www.shizukalab.com/toolkits/sna/triangle-transitivity
}

\examples{
ttri(caribou)
}
