% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata keywords to a file using exifTools}
\usage{
add_metadata(filePath, exifPath = "../../../exiftool-12.65/exiftool.exe")
}
\arguments{
\item{filePath}{Which file to add metadata to?}

\item{exifPath}{The filepath to the ExifTools executable.}
}
\description{
A way to document which params were used for a figure, and to
track the file and latest git commit. Only tested with the windows
executable.
}
