% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_metadata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata keywords to a pdf using exifTools}
\usage{
add_metadata(
  filePath,
  ...,
  incGit = TRUE,
  incSession = FALSE,
  exifPath = "../../../exiftool-12.65/exiftool.exe"
)
}
\arguments{
\item{filePath}{Which file to add metadata to? Will fail if not a ".pdf".
In future, more options may be added if that makes sense.}

\item{...}{Any additional information that needs to be included beyond the
params and file code? Include as '"name" =  value'.}

\item{incGit}{BINARY: Should the full output from \code{\link[=get_latest_git]{get_latest_git()}} be
included? Even if FALSE, the hash of the last commit is still included in
the file code.}

\item{incSession}{BINARY: Should the output from \code{\link[=sessionInfo]{sessionInfo()}} be
included? This is currently not implemented as it looks more complex than
I had anticipated.}

\item{exifPath}{The file path to the ExifTools executable.}
}
\description{
A way to help document which params were used for a figure, to
track which file it was produced from, and the last git commit before it
was created, plus any other information we want to add. Only tested with
the windows executable exiftools. Only works for pdf files right now.
}
