% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat2.R
\name{cat2}
\alias{cat2}
\title{Concatenate the name and variable}
\usage{
cat2(x, show = TRUE)
}
\arguments{
\item{x}{The variable that should be displayed. It's name and value will be
shown. Lists are unlisted. Not tested with complex data types - use with
caution! If you just want the variable (and not the name), stick with
\code{cat()}, or use \code{cat3()}, which basically adds "\n" before and after x.}

\item{show}{logical If FALSE, the function does nothing; it is useful if
there are multiple \code{cat2()} calls throughout; set a variable at the
beginning and use that as the 'show' argument in each \code{cat2()} call.
Simply changing that to FALSE allows the function to be run without all
the calls, but the calls can remain in place if further \code{cat2()} use is
likely.}
}
\description{
This function is likely useful because I don't know how to
debug "correctly". When developing functions, it can be hard to know
where the code is going wrong. Using \code{cat2()} prints out a variable and
it's name. Sprinkle these at key locations in functions to see what is
happening where.
}
