% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_list.R
\name{cat_list}
\alias{cat_list}
\title{Print out a list in a slightly prettier way (in my opinion!)}
\usage{
cat_list(
  list,
  name = NULL,
  listDepth = 0,
  nameLength = 1,
  maxDepth = 1,
  verbose = TRUE,
  show = TRUE
)
}
\arguments{
\item{list}{The list to be printed out.}

\item{name}{Name to be printed at the top of each list. If NULL (the
default), it matches the name of the 'list' argument. This is mainly here
for recursion so the "iiData" in the loop isn't used as the name for
every subsequent nested list.}

\item{listDepth}{Used for recursive calls to track the depth; each
subsequent depth is marked by an extra "O" at the start of the line.}

\item{nameLength}{Used for recursion to help things align.}

\item{maxDepth}{Used for recursion to help things align.}

\item{verbose}{If TRUE (the default), data is shown in full; if FALSE, the
dimensions of each item in the list are shown.}

\item{show}{LOGICAL: If FALSE, the function does nothing; it is useful if
there are multiple \code{cat_list()} calls throughout; set a variable at the
beginning and use that as the 'show' argument in each \code{cat_list()} call.
Simply changing that to FALSE allows the function to be run without all
the calls, but the calls can remain in place if further \code{cat_list()} use
is likely.}
}
\description{
Printing a list doesn't look great, and \code{cat()} doesn't work.
This function is an alternative way to show the data in a list. It is
still in progress, and currently only handles vectors, matrices and
lists/data.frames.
}
