% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_list_names.R
\name{get_list_names}
\alias{get_list_names}
\title{Get the names of all values in nested lists}
\usage{
get_list_names(list)
}
\arguments{
\item{list}{The nested list.}
}
\description{
This function is basically a recursive calling of "names" to
get the names of everything in nested lists.
}
