% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Prepare metadata about a file}
\usage{
get_metadata(format = "named", ..., incGit = TRUE, incSession = TRUE)
}
\arguments{
\item{format}{Which format do you want the metadata in? Options are "pdf",
"NetCDF", "print", "yaml" or "json" or "list" (all the same), or "named".
"named" is the default and the fallback if anything except one of these
strings is entered (cASe inSENSItIve).

Options (examples of formatting):

\if{html}{\out{<div class="sourceCode">}}\preformatted{ "pdf"                      "-Keywords=month:Dec"
 "NetCDF"                   "month=Dec"
 "print"                    "month : Dec"
 "yaml" / "json" / "list"   list("month" = "Dec")
 "named"                    "month" = "Dec"
}\if{html}{\out{</div>}}}

\item{...}{Any additional information that needs to be included beyond the
params and file code? Include as '"name" =  value'.}

\item{incGit}{BINARY: Should the full output from \code{\link[=get_latest_git]{get_latest_git()}} be
included? Even if FALSE, the hash of the last commit is still included.}

\item{incSession}{BINARY: Should the output from \code{\link[=sessionInfo]{sessionInfo()}} be
included? This is currently not implemented as it looks more complex than
I had anticipated.}
}
\description{
Get the information necessary for documenting how a file /
dataset was created from a .qmd notebook. It only works within a .qmd
notebook as it directly looks for the 'params' defined at the top.
}
