% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now.R
\name{now}
\alias{now}
\title{Return the formatted current date and/or time quicker}
\usage{
now(format = "F")
}
\arguments{
\item{format}{"string": Which format to return? See \code{strptime()}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{The options are:
    "n"    "yyyy-mm-dd HH:MM:ss zzz"     e.g. "2023-07-29 15:24:43 AEST
    "F"    "yyyy-mm-dd"                  e.g. "2023-07-29"
    "t"    "HH:MM:ss zzz"                e.g. "15:24:43"
    "Y"    yyyy                          e.g. 2023
    "m"    m                             e.g. 7
    "b"    "Mth"                         e.g. "Jul"
    "B"    "Month"                       e.g. "July"
    "d"    dd                            e.g. 29
    ""     anything else accepted by `strptime()`
}\if{html}{\out{</div>}}}
}
\description{
When we want the current date and/or time, we need awkward
and/or verbose syntax to get the correct formatting. This function offers
some simpler options. Basically just a wrapper around
\verb{format(Sys.time(), \%\%)}, but "n" and "t" behave differently.
}
