% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_line.R
\name{print_line}
\alias{print_line}
\title{Print a line to the console to help delineate output}
\usage{
print_line(lineType = "=", nStart = 1, nEnd = 1, width = 22, show = TRUE)
}
\arguments{
\item{lineType}{"string": The symbol/s that will be printed to create a
line across the console. A numeric value between 1 and 9 can be used,
with the following defined:

Numeric Options

\if{html}{\out{<div class="sourceCode">}}\preformatted{    1      ---------------- 1
    2      ================ 2
    3      ________________ 3
    4      ~~~~~~~~~~~~~~~~ 4
    5      ++++++++++++++++ 5
    6      >>>>>>>>>>>>>>>> 6
    7      <<<<<<<<<<<<<<<< 7
    8      |||||||||||||||| 8
    9      !!!!!!!!!!!!!!!! 9
}\if{html}{\out{</div>}}

All other numbers will default to "-" (equivalent to 1). To use a line of
the number itself, enclose it as a string: "1". See example.}

\item{nStart}{numeric: How many \verb{\\n} should be included \emph{before} the line?}

\item{nEnd}{numeric: How many \verb{\\n} should be included \emph{after} the line?}

\item{width}{numeric: How many times should the 'lineType' be repeated?
Usually unnecessary, but change it if screen is particularly wide or
narrow.}

\item{show}{LOGICAL; should the line be printed? Can be set as FALSE along
with \code{\link[=cat2]{cat2()}} so it can be easily switched on or off for debugging.}
}
\description{
Useful when working interactively in RStudio, and can be
particularly helpful to group output to the console or when testing.
}
\examples{
  print_line()
  print_line(1)
  print_line("1")
  print_line(2)
  print_line("=", 1, 0)
  print_line("-", width = 25)
  print_line("?", width = 10)
  print_line("+-", width = 10)

}
