% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_list.R
\name{save_list}
\alias{save_list}
\title{A very simple wrapper around \code{\link[jsonlite:read_json]{jsonlite::write_json()}}}
\usage{
save_list(x, filename)
}
\arguments{
\item{x}{The list to save.}

\item{filename}{The filename (and path relative to the working directory)
indicating where to save the list. Must included the file extension.}
}
\description{
This function will be expanded to also accommodate yaml, rds
and rData output, and different json options. It currently just converts
an R list to a json format, and saves it. It uses the default values of
'pretty' = TRUE, and 'dataframe' = "columns".
}
