% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timers.R
\name{start_timer}
\alias{start_timer}
\title{Starts the timer}
\usage{
start_timer()
}
\description{
This function and \code{\link[=end_timer]{end_timer()}} are mainly because I can never
remember the syntax/name of \code{\link[=proc.time]{proc.time()}}. Simply call \code{\link[=start_timer]{start_timer()}}
before the process you are timing begins, and then call \code{\link[=end_timer]{end_timer()}}
afterwards, feeding it with the output of \code{\link[=start_timer]{start_timer()}}.
}
\examples{
\dontrun{
  start <- start_timer()
  Sys.sleep(0.2) # do code!
  end_timer(start)
}
}
